+++ libsel4allocman/include/allocman/mspace/k_r_malloc.h
@@ -11,14 +11,14 @@
 #ifndef _ALLOCMAN_MSPACE_K_R_MALLOC_H_
 #define _ALLOCMAN_MSPACE_K_R_MALLOC_H_
 
-#include <stdint.h>
+#include <stdlib.h>
 
 /* A K&R malloc style allocation that can be 'put in a box' as it were. */
 
 typedef union k_r_malloc_header {
     struct {
         union k_r_malloc_header *ptr;
-        uint32_t size;
+        size_t size;
     } s;
     /* Force alignment */
     long long x;
@@ -27,12 +27,12 @@ typedef union k_r_malloc_header {
 typedef struct mspace_k_r_malloc {
     k_r_malloc_header_t base;
     k_r_malloc_header_t *freep;
-    uint32_t cookie;
-    k_r_malloc_header_t *(*morecore)(uint32_t cookie, struct mspace_k_r_malloc *k_r_malloc, uint32_t new_units);
+    size_t cookie;
+    k_r_malloc_header_t *(*morecore)(size_t cookie, struct mspace_k_r_malloc *k_r_malloc, size_t new_units);
 } mspace_k_r_malloc_t;
 
-void mspace_k_r_malloc_init(mspace_k_r_malloc_t *k_r_malloc, uint32_t cookie, k_r_malloc_header_t * (*morecore)(uint32_t cookie, mspace_k_r_malloc_t *k_r_malloc, uint32_t new_units));
-void *mspace_k_r_malloc_alloc(mspace_k_r_malloc_t *k_r_malloc, uint32_t nbytes);
+void mspace_k_r_malloc_init(mspace_k_r_malloc_t *k_r_malloc, size_t cookie, k_r_malloc_header_t * (*morecore)(size_t cookie, mspace_k_r_malloc_t *k_r_malloc, size_t new_units));
+void *mspace_k_r_malloc_alloc(mspace_k_r_malloc_t *k_r_malloc, size_t nbytes);
 void mspace_k_r_malloc_free(mspace_k_r_malloc_t *k_r_malloc, void *ap);
 
 #endif
