+++ libsel4allocman/include/allocman/mspace/virtual_pool.h
@@ -12,7 +12,7 @@
 #define _ALLOCMAN_MSPACE_VIRTUAL_POOL_H_
 
 #include <autoconf.h>
-#include <stdint.h>
+#include <stdlib.h>
 #include <sel4/types.h>
 #include <allocman/mspace/mspace.h>
 #include <allocman/mspace/k_r_malloc.h>
@@ -21,7 +21,7 @@
 
 struct mspace_virtual_pool_config {
     void *vstart;
-    uint32_t size;
+    size_t size;
     seL4_CPtr pd;
 };
 
@@ -36,8 +36,8 @@ typedef struct mspace_virtual_pool {
 
 void mspace_virtual_pool_create(mspace_virtual_pool_t *virtual_pool, struct mspace_virtual_pool_config config);
 
-void *_mspace_virtual_pool_alloc(struct allocman *alloc, void *_virtual_pool, uint32_t bytes, int *error);
-void _mspace_virtual_pool_free(struct allocman *alloc, void *_virtual_pool, void *ptr, uint32_t bytes);
+void *_mspace_virtual_pool_alloc(struct allocman *alloc, void *_virtual_pool, size_t bytes, int *error);
+void _mspace_virtual_pool_free(struct allocman *alloc, void *_virtual_pool, void *ptr, size_t bytes);
 
 static inline struct mspace_interface mspace_virtual_pool_make_interface(mspace_virtual_pool_t *virtual_pool) {
     return (struct mspace_interface){
