+++ libsel4allocman/src/mspace/dual_pool.c
@@ -16,8 +16,8 @@
 void mspace_dual_pool_create(mspace_dual_pool_t *dual_pool, struct mspace_fixed_pool_config config)
 {
     *dual_pool = (mspace_dual_pool_t) {
-        .fixed_pool_start = (uint32_t)config.pool,
-        .fixed_pool_end = (uint32_t)config.pool + config.size,
+        .fixed_pool_start = (uintptr_t)config.pool,
+        .fixed_pool_end = (size_t)config.pool + config.size,
         .have_virtual_pool = 0
     };
     mspace_fixed_pool_create(&dual_pool->fixed_pool, config);
@@ -28,7 +28,7 @@ void mspace_dual_pool_attach_virtual(mspace_dual_pool_t *dual_pool, struct mspac
     dual_pool->have_virtual_pool = 1;
 }
 
-void *_mspace_dual_pool_alloc(allocman_t *alloc, void *_dual_pool, uint32_t bytes, int *error)
+void *_mspace_dual_pool_alloc(allocman_t *alloc, void *_dual_pool, size_t bytes, int *error)
 {
     mspace_dual_pool_t *dual_pool = (mspace_dual_pool_t*)_dual_pool;
     /* if we have a virtual pool try and use this instead of wasting time in the fixed pool */
@@ -44,10 +44,10 @@ void *_mspace_dual_pool_alloc(allocman_t *alloc, void *_dual_pool, uint32_t byte
     return _mspace_fixed_pool_alloc(alloc, &dual_pool->fixed_pool, bytes, error);
 }
 
-void _mspace_dual_pool_free(struct allocman *alloc, void *_dual_pool, void *ptr, uint32_t bytes)
+void _mspace_dual_pool_free(struct allocman *alloc, void *_dual_pool, void *ptr, size_t bytes)
 {
     mspace_dual_pool_t *dual_pool = (mspace_dual_pool_t*)_dual_pool;
-    if ( (uint32_t)ptr >= dual_pool->fixed_pool_start && (uint32_t)ptr < dual_pool->fixed_pool_end) {
+    if ( (uintptr_t)ptr >= dual_pool->fixed_pool_start && (uintptr_t)ptr < dual_pool->fixed_pool_end) {
         _mspace_fixed_pool_free(alloc, &dual_pool->fixed_pool, ptr, bytes);
     } else {
         _mspace_virtual_pool_free(alloc, &dual_pool->virtual_pool, ptr, bytes);
